Save and cancel order adapted to components integration

+++ java/mtsj/core/src/main/java/io/oasp/application/mtsj/ordermanagement/logic/impl/OrdermanagementImpl.java
@@ -1,5 +1,9 @@
 package io.oasp.application.mtsj.ordermanagement.logic.impl;
 
+import java.sql.Timestamp;
+import java.time.Instant;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
 
 import javax.inject.Inject;
@@ -7,11 +11,22 @@ import javax.inject.Named;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.transaction.annotation.Transactional;
 
+import io.oasp.application.mtsj.bookingmanagement.common.api.datatype.BookingType;
+import io.oasp.application.mtsj.bookingmanagement.logic.api.Bookingmanagement;
+import io.oasp.application.mtsj.bookingmanagement.logic.api.to.BookingCto;
 import io.oasp.application.mtsj.bookingmanagement.logic.api.to.BookingEto;
+import io.oasp.application.mtsj.bookingmanagement.logic.api.to.BookingSearchCriteriaTo;
 import io.oasp.application.mtsj.bookingmanagement.logic.api.to.InvitedGuestEto;
+import io.oasp.application.mtsj.bookingmanagement.logic.api.to.InvitedGuestSearchCriteriaTo;
 import io.oasp.application.mtsj.general.logic.base.AbstractComponentFacade;
+import io.oasp.application.mtsj.ordermanagement.common.api.exception.CancelNotAllowedException;
+import io.oasp.application.mtsj.ordermanagement.common.api.exception.NoBookingException;
+import io.oasp.application.mtsj.ordermanagement.common.api.exception.NoInviteException;
+import io.oasp.application.mtsj.ordermanagement.common.api.exception.OrderAlreadyExistException;
+import io.oasp.application.mtsj.ordermanagement.common.api.exception.WrongTokenException;
 import io.oasp.application.mtsj.ordermanagement.dataaccess.api.OrderEntity;
 import io.oasp.application.mtsj.ordermanagement.dataaccess.api.OrderLineEntity;
 import io.oasp.application.mtsj.ordermanagement.dataaccess.api.dao.OrderDao;
@@ -19,6 +34,7 @@ import io.oasp.application.mtsj.ordermanagement.dataaccess.api.dao.OrderLineDao;
 import io.oasp.application.mtsj.ordermanagement.logic.api.Ordermanagement;
 import io.oasp.application.mtsj.ordermanagement.logic.api.to.OrderCto;
 import io.oasp.application.mtsj.ordermanagement.logic.api.to.OrderEto;
+import io.oasp.application.mtsj.ordermanagement.logic.api.to.OrderLineCto;
 import io.oasp.application.mtsj.ordermanagement.logic.api.to.OrderLineEto;
 import io.oasp.application.mtsj.ordermanagement.logic.api.to.OrderLineSearchCriteriaTo;
 import io.oasp.application.mtsj.ordermanagement.logic.api.to.OrderSearchCriteriaTo;
@@ -48,6 +64,12 @@ public class OrdermanagementImpl extends AbstractComponentFacade implements Orde
   @Inject
   private OrderLineDao orderLineDao;
 
+  @Inject
+  private Bookingmanagement bookingManagement;
+
+  @Value("${mythaistar.hourslimitcancellation}")
+  private int hoursLimit;
+
   /**
    * The constructor.
    */
@@ -64,7 +86,7 @@ public class OrdermanagementImpl extends AbstractComponentFacade implements Orde
     OrderCto cto = new OrderCto();
     cto.setBooking(getBeanMapper().map(entity.getBooking(), BookingEto.class));
     cto.setHost(getBeanMapper().map(entity.getHost(), BookingEto.class));
-    cto.setOrderLines(getBeanMapper().mapList(entity.getOrderLines(), OrderLineEto.class));
+    cto.setOrderLines(getBeanMapper().mapList(entity.getOrderLines(), OrderLineCto.class));
     cto.setOrder(getBeanMapper().map(entity, OrderEto.class));
     cto.setInvitedGuest(getBeanMapper().map(entity.getInvitedGuest(), InvitedGuestEto.class));
     return cto;
@@ -82,22 +104,52 @@ public class OrdermanagementImpl extends AbstractComponentFacade implements Orde
   public boolean deleteOrder(Long orderId) {
 
     OrderEntity order = getOrderDao().find(orderId);
+
+    if (!cancellationAllowed(order)) {
+      throw new CancelNotAllowedException();
+    }
+    OrderLineSearchCriteriaTo criteria = new OrderLineSearchCriteriaTo();
+    criteria.setOrderId(order.getId());
+    List<OrderLineEntity> orderLines = this.orderLineDao.findOrderLines(criteria).getResult();
+
+    for (OrderLineEntity orderLine : orderLines) {
+      this.orderLineDao.delete(orderLine.getId());
+    }
     getOrderDao().delete(order);
     LOG.debug("The order with id '{}' has been deleted.", orderId);
     return true;
   }
 
   @Override
-  public OrderEto saveOrder(OrderEto order) {
+  public OrderEto saveOrder(OrderCto order) {
 
     Objects.requireNonNull(order, "order");
+    List<OrderLineCto> linesCto = order.getOrderLines();
+    List<OrderLineEntity> orderLineEntities = new ArrayList<>();
+    for (OrderLineCto lineCto : linesCto) {
+      OrderLineEntity orderLineEntity = getBeanMapper().map(lineCto, OrderLineEntity.class);
+      orderLineEntity.setExtras(lineCto.getExtras());
+      orderLineEntity.setDishId(lineCto.getOrderLine().getDishId());
+      orderLineEntity.setAmount(lineCto.getOrderLine().getAmount());
+      orderLineEntity.setComment(lineCto.getOrderLine().getComment());
+      orderLineEntities.add(orderLineEntity);
+    }
+
     OrderEntity orderEntity = getBeanMapper().map(order, OrderEntity.class);
 
     // initialize, validate orderEntity here if necessary
-    OrderEntity resultEntity = getOrderDao().save(orderEntity);
-    LOG.debug("Order with id '{}' has been created.", resultEntity.getId());
+    orderEntity = getValidatedOrder(orderEntity.getBooking().getBookingToken(), orderEntity);
+    orderEntity.setOrderLines(orderLineEntities);
+    OrderEntity resultOrderEntity = getOrderDao().save(orderEntity);
+    LOG.debug("Order with id '{}' has been created.", resultOrderEntity.getId());
 
-    return getBeanMapper().map(resultEntity, OrderEto.class);
+    for (OrderLineEntity orderLineEntity : orderLineEntities) {
+      orderLineEntity.setOrderId(resultOrderEntity.getId());
+      OrderLineEntity resultOrderLine = getOrderLineDao().save(orderLineEntity);
+      LOG.info("OrderLine with id '{}' has been created.", resultOrderLine.getId());
+    }
+
+    return getBeanMapper().map(resultOrderEntity, OrderEto.class);
   }
 
   /**
@@ -118,11 +170,18 @@ public class OrdermanagementImpl extends AbstractComponentFacade implements Orde
   }
 
   @Override
-  public PaginatedListTo<OrderLineEto> findOrderLineEtos(OrderLineSearchCriteriaTo criteria) {
+  public PaginatedListTo<OrderLineCto> findOrderLineCtos(OrderLineSearchCriteriaTo criteria) {
 
     criteria.limitMaximumPageSize(MAXIMUM_HIT_LIMIT);
     PaginatedListTo<OrderLineEntity> orderlines = getOrderLineDao().findOrderLines(criteria);
-    return mapPaginatedEntityList(orderlines, OrderLineEto.class);
+    List<OrderLineCto> orderLinesCto = new ArrayList<>();
+    for (OrderLineEntity orderline : orderlines.getResult()) {
+      OrderLineCto orderLineCto = new OrderLineCto();
+      orderLineCto.setOrderLine(getBeanMapper().map(this.orderLineDao.findOne(orderline.getId()), OrderLineEto.class));
+      orderLineCto.setExtras(orderline.getExtras());
+      orderLinesCto.add(orderLineCto);
+    }
+    return new PaginatedListTo<>(orderLinesCto, orderlines.getPagination());
   }
 
   @Override
@@ -157,4 +216,87 @@ public class OrdermanagementImpl extends AbstractComponentFacade implements Orde
     return this.orderLineDao;
   }
 
+  private OrderEntity getValidatedOrder(String token, OrderEntity orderEntity) {
+
+    // BOOKING VALIDATION
+    if (getOrderType(token) == BookingType.COMMON) {
+      BookingEto booking = getBooking(token);
+      if (booking == null) {
+        throw new NoBookingException();
+      }
+      List<OrderEto> currentOrders = getBookingOrders(booking.getId());
+      if (!currentOrders.isEmpty()) {
+        throw new OrderAlreadyExistException();
+      }
+      orderEntity.setBookingId(booking.getId());
+
+      // GUEST VALIDATION
+    } else if (getOrderType(token) == BookingType.INVITED) {
+
+      InvitedGuestEto guest = getInvitedGuest(token);
+      if (guest == null) {
+        throw new NoInviteException();
+      }
+      List<OrderEto> currentGuestOrders = getInvitedGuestOrders(guest.getId());
+      if (!currentGuestOrders.isEmpty()) {
+        throw new OrderAlreadyExistException();
+      }
+      orderEntity.setBookingId(guest.getBookingId());
+      orderEntity.setInvitedGuestId(guest.getId());
+    }
+
+    return orderEntity;
+  }
+
+  private BookingType getOrderType(String token) throws WrongTokenException {
+
+    if (token.startsWith("CB_")) {
+      return BookingType.COMMON;
+    } else if (token.startsWith("GB_")) {
+      return BookingType.INVITED;
+    } else {
+      throw new WrongTokenException();
+    }
+  }
+
+  private BookingEto getBooking(String token) {
+
+    BookingSearchCriteriaTo criteria = new BookingSearchCriteriaTo();
+    criteria.setBookingToken(token);
+    PaginatedListTo<BookingEto> booking = this.bookingManagement.findBookingEtos(criteria);
+    return booking.getResult().isEmpty() ? null : booking.getResult().get(0);
+  }
+
+  private List<OrderEto> getBookingOrders(Long idBooking) {
+
+    OrderSearchCriteriaTo criteria = new OrderSearchCriteriaTo();
+    criteria.setBookingId(idBooking);
+    return findOrderEtos(criteria).getResult();
+  }
+
+  private InvitedGuestEto getInvitedGuest(String token) {
+
+    InvitedGuestSearchCriteriaTo criteria = new InvitedGuestSearchCriteriaTo();
+    criteria.setGuestToken(token);
+    PaginatedListTo<InvitedGuestEto> guest = this.bookingManagement.findInvitedGuestEtos(criteria);
+    return guest.getResult().isEmpty() ? null : guest.getResult().get(0);
+  }
+
+  private List<OrderEto> getInvitedGuestOrders(Long idInvitedGuest) {
+
+    OrderSearchCriteriaTo criteria = new OrderSearchCriteriaTo();
+    criteria.setInvitedGuestId(idInvitedGuest);
+    return findOrderEtos(criteria).getResult();
+  }
+
+  private boolean cancellationAllowed(OrderEntity order) {
+
+    BookingCto booking = this.bookingManagement.findBooking(order.getBookingId());
+    Timestamp bookingTime = booking.getBooking().getBookingDate();
+    Long bookingTimeMillis = bookingTime.getTime();
+    Long cancellationLimit = bookingTimeMillis - (3600000 * this.hoursLimit);
+    Long now = Timestamp.from(Instant.now()).getTime();
+
+    return (now > cancellationLimit) ? false : true;
+  }
 }
